Initial config/schema scaffold for the host vSwitch resource
This is an attempt to properly map out the complex structure of the data
objects that are involved in working with the host vSwitch management
tasks.

This is the kind of pattern that I want to use going forward in the
provider. In fact, it's entirely possible that some of these schema and
conversion functions will be used in other resources as well, as some
other management objects re-use different data objects, so having all of
these mapped out is useful and will ultimately introduce consistency in
how the schemas are structured across resources.