Add test case for VM refresh panic
This adds a test case for @mjrider's new network device polling logic
that fixes several issues resulting from the inconsistencies between
physical devices and IP information when the guest is shut down. The
test is similar to the VM test, but in a second step it powers down the
VM and does another plan. This effectively triggers the panic on the old
code.