r/host_virtual_switch: Finish basic tests, work on diff issues
Using such a complex nested set is causing issues with diffs that
probably will not be able to be solved by setting computed values,
namely because there is not much of a way we will be able to rely on the
hashes on the sets being predictable.

I want to convert this to use lists all the way down, which may fix this
on part of the fact that there are no hashes involved, which should
remove a big factor of issues with diffing this.

Preserving the nested structure of this resource is important, as this
will give us a template for properly working with and visualizing the
vSphere API in the future without a massive translation of data from a
flattened, truncated form to another (with this nested setup, nearly all
attributes are exposed and located where they would be located in the
vmomi managed and data object hierarchy).