r/host_virtual_switch: Move resource schema to flat layout
Constructing a host vSwitch spec is a very complex operation, and laying
out the schema in a series of complex nested sets has proven to be more
trouble than it's worth - a lack of solid upstream default generation
for nested resources, ambiguity of what it means if a specific resource
structure is missing, and the calculation and saving of computed values
makes it very hard to use this kind of structure with the vSphere API.

As such, I've moved the schema for the host vSwitch to an almost
entirely flat setup, and populated almost the entire spec with defaults.
The resulting set actually makes configs look a bit simpler in TF code,
and we can still deeply populate the entire spec in a functional manner,
in addition to reading from it.

All tests now work - the only missing thing after fixing all of this was
to add a few fields that the vSphere API docs say are deprecated and not
used, but are still showing up in the API anyway.