r/host_virtual_switch: Make all NIC parameters mandatory
The computed nature of the default NIC failover order in the policy was
causing issues when removing NICs - since the user had not defined the
policy, the old computed order was being sent to server.

There are too many scenarios with NIC order to really have a sane
default here. A NIC that is bound to a vSwitch can be in the active
list, it can be in the standby list, or it can be in neither and just be
unused. As such, we can't reasonably assume what the failover order is
here, so the user really needs to explicitly state it.

We will also need to figure out what to do about inheritance on port
groups as a nil failover order is what denotes inheritance here. We may
have to add a flag, but I would personally like to avoid that.