r/host_virtual_switch: Move NIC order requirement to transform
This *should* set us up to be able to allow these fields to be computed
in the port group resource, in addition to being able to allow the state
of both of these fields to be unset, to indicate inheritance. To help
with this I've moved these fields to have their values fetched with
GetOkExists.