New data source: vsphere_host
This resource can be used to discover the managed object ID of a
specific host. Like vsphere_datacenter, this has been written to
facilitate the use of a managed object reference ID to fetch the host
system object to work with various resources, such as host networking
resources or datastores.

Made the function name more terse as well.