New data source: vsphere_vmfs_disks
This adds the vsphere_vmfs_disks data source, which allows someone to
discover storage devices that can be used for VMFS datastores on a given
ESXi host.

The output of this is intended to be fed into a vsphere_vmfs_datastore
resource, which allows for the creation of a VMFS datastore off a host
or a set of hosts (ie: in the case of FC or iSCSI disks).