r/vmfs_datastore: Better error handling
Moved helpers to check to see if removed datastores were missing over to
more general and type-safe VIM fault helpers (introduced in
vim_helper.go). This is how we will be checking for errors in the future
- a bit of back-refactoring may need to be done later to accommodate
this in existing code, but the path to check should be a lot more clear
now.