r/vmfs_datastore: Add folder support
Just as a bonus, added folder support. This commit also ultimately
scaffolds a lot of stuff that we will need for proper inventory support
anyway, including helpers to determine well-known paths for the 4 major
inventory types based off an existing managed object, which will allow
for some saner control around where in inventory a specific item goes.
As an example, with this approach, the vsphere_vmfs_datastore resource
"folder" attribute can only place datastores in folders that are in the
datacenter that the datastore is actually being set up in - so if the
host that the datastore is being set up in is in the "dc1" datacenter,
the folder needs to exist within "/dc1/datastore".