r/vmfs_datastore: Make constants more specific, remove update state save
The constant names were really general and we may have run into
collisions eventually. Also, we don't need to record every disk update
like I thought for some reason - if there is an error, it will just be
picked up next refresh anyway. Removed this extra bit of complexity to
be consistent with the fact that we really don't do this anywhere else.