r/vmfs_datastore: Fix import
Smoke testing the import uncovered issues where certain fields weren't
populated. Unfortunately populating all the fields properly and
validating that the datastore is actually a datastore that the host has
mounted also means that we need to know the host ID, so we have had to
add that to the import ID.

This has the unfortunate side effect of making it hard to test this,
currently. I have created hashicorp/terraform#16036 to help address
this, and will re-enable the test when it gets merged.