r/virtual_machine: Enforce powered on state
This commit adds something that we generally assume is correct in
Terraform - that an instance's power state is either one of two states:
on, or gone.

vsphere_virtual_machine has had a couple of issues that have been
attributed to power state in the last few months that have been fixed,
but the issue of ensuring that VMs are always powered on has not been
touched on yet.

In this commit, we add the `power_state` pseudo-computed variable,
working around the fact that the provider can't manipulate the diff
directly by using an immutable default to always indicate that a VM that
is powered off or suspended will be turned back on during the next
Terraform run. If a VM is powered off, this turns the VM back on during
the configuration modification phase at the end of the update.

Fixes #134.
Related to #103.