website: Add resource categories
We are getting more resources now, and it will be worthwhile to start
departing them into categories now.

I am breaking down the categories mainly into their specific resource
classes as they are generally seen in vCenter:

* Hosts and Clusters
* VMs and Templates
* Storage
* Networking

There are other classes as well that will have resources that are not
necessarily related to be above that make logical sense:

* Administration - For vCenter and vSphere SSO related admin resources.
Licenses are in this category currently and items like tags and what not
will probably fall here as well.
* Inventory - For inventory management resources, such as datacenters
(which is a fancy top-level resource grouping) and folders (which can be
used across all four of the major resource classes).