r/virtual_machine: Wait for customization properly, and routed net
This is a two-part change that ultimately changes/fixes customization
workflow, but has some other implications:

* First off, we are now waiting for customization to complete by
watching VM events. This is the correct way to wait for customization to
complete as the actual customization task returns very quickly, while
the actual customization process can actually take a few minutes and
fail out-of-band of the initial task.

* Second off, we are now no longer waiting for all interfaces to have IP
addresses during Read, rather we are now waiting for a **routeable**
network. This fixes situations where a template may be configured for
DHCP and gets an autoconfiguration address before actually getting an
IP. This seems to come up during Windows customization especially,
which is what we are trying to fix with this work, so having the event
watch fix without having this one would not fix the issue completely.

* The `wait_for_guest_net` option has also been added, which allows
someone to completely turn off the network waiter. This should help
alleviate some edge cases where NICs have not been configured with IP
addresses, and also allows someone to bypass this behaviour if they are
not configuring a gateway on the VM through either static configuration
of DHCP.

Fixes #140.