r/virtual_machine: Refactor tests
Maintaining this file has become increasingly difficult. As such, I have
re-factored the tests in the style as we are doing most of the new tests
- highly functional, static configs with a tolerable level of
repetition, and all tests set up as subtests, so every test is easy to
locate as part of the larger test case list.

I have also removed a few tests that currently do not function, for one
reason or another. A couple of these are due to present bugs in the
resource, but as pending work may deprecate the resource in its current
form, I think maintenance effort is better spent elsewhere as they are
not currently reported business critical bugs that are rendering the
resource unusable.

The tests were:

* A test with a CDROM virtual device - this test was panicking. As it's
hard to imagine who is using VMs in TF with CDROM devices attached, I
think this is a bit of an edge case.
* A test for a NIC with DHCP. In theory this should work, but it's hard
to test properly without dealing with races in VMware tools, so the test
itself probably did not have much value.
* A test to detach disks added out-of-band. This is indeed a bug as Read
does not seem to be picking up the extra disk and triggering the needed
diff. Again, this is something that will be picked up on the refactor,
or might look completely different afterwards.

Some other tests were consolidated - memory and CPU changes are now one
test, and the individual keep_on_remove test for disk devices was
dropped as we now test this with the test that adds an external disk via
vsphere_virtual_disk, where a lack of keep_on_remove on the attachment
actually causes issues during destroy as the disk is removed
prematurely, with the virtual machine.