r/virtual_machine: Tests for Windows, also move net waiter
Adding in the test for all of the customization work we have done, which
mainly crops up with a Windows template.

Also added a new test for DHCP only without waiting on guest networking.
Triggering this test to fail deliberately uncovered an issue with how
the network waiter basically prevents destroy of powered on virtual
machines when we don't have routeable guest network info, so I have
finally just moved the net waiter to Create and Read, which is a more
suitable place for a process that waits for a VM to be ready anyway.