New resource: vsphere_tag_category
This adds a new resource, vsphere_tag_category, which is the start of
the effort to get tag support into the provider. This allows one to
create and manage tag categories in the provider, which can then be
passed along to tag resources.

As part of this work, other work has been done to prime the provider for
the ability to manage multiple API connections at once, as the provider
now needs to connect to two to do its job. Furthermore, tags are not
supported on all versions of vSphere and not on ESXi, so extra
precautions needed to be undertaken to ensure these cases were properly
blocked.