r/tag_category: Drop used_by from the attribute list
I can't find any sort of useful data on how this is used, the docs don't
mention it, the API is murky on its purpose and the UI does not expose
it. Adding a tag to an object with the category doesn't add anything to
it either. Since we can't figure out what it's used for, I'm just
dropping it - if it's missed by any user, we can just add it back later.