r/tag_category: Add a "this is a bug" error message to import
For the "multiple category returned" case. By all observations this is a
case that can't happen, but VIC's tag package seems to think it can, so
we are indicating that the user needs to put in an issue if they indeed
somehow managed to add multiple tag categories with the same name.