r/virtual_machine: Increase customization timeout to 10 mins
Users have been reporting that 5 minutes is probably too short. Indeed,
in our lab it takes about 3 mins to complete (using nested ESXi and
datastores that are on SSDs), so this is definitely reasonable.

Adjusting this to 10 minutes which should hopefully be good for the vast
majority of cases. If need be, we can make this configurable, but then
we are wading into provider timeout territory as well.