New resource: vsphere_tag
This adds the vsphere_tag resource, which can be used to set up and
manage tags. Data source and docs will be coming soon.

Note that the import for this function takes a JSON string, as both tags
and categories can support special characters, so there is not much for
us to really use as a delimiter.