tags: Support tagging on vsphere_virtual_machine
This commit contains most of the up-front work necessary to tag objects,
in addition to adding tagging capability to vsphere_virtual_machine.
With the work done here, adding tagging support to any resource that can
support it should be a trivial act.

The one major exception is vsphere_host_port_group, which is an odd case
as created port groups appear to be aggregated into a single network in
vSphere. This commit includes helpers to determine a Network object ID
from a specific network name and HostSystem, but are not tested at this
point in time. We will probably need to support this case via a
vsphere_host_network_tag_association resource and a
vsphere_standard_network data source, the latter to derive the correct
network MO and the former to associate the tag correctly. These items
will be added once tagging is added for all other objects, so that they
can be added in a different PR.