tags: Support tagging in datastore resources
This update supports tagging in the vsphere_nas_datastore and
vsphere_vmfs_datastore resources. Using the same workflows applied to
the virtual machine resource.

In addition to the tests supplied for the datastore resources, I've
re-enabled the import test for vsphere_vmfs_datastore now that we have
ImportStateIdFunc to help generate the ID.