r/virtual_machine: Network fixes
Along with 1f4c85f, this corrects a couple of issues and adds tests for
one that we can test for:

Fixes #165, which was an issue where network searches in govmomi were
not working for networks with unescaped slashes in their names. We also
have a test for this now as well where we reproduce by adding a
host-level portgroup, which the API does not escape slashes for.

Fixes #161, which was an issue with network searches where panics could
happen on trying to load a network backed by a 3rd party hardware DVS.
The latest govmomi has commits that properly use base types for the
respective items.