r/folder: Fix state migration
My understanding of the MigrateState API was off, as it seems that the
return value for StateMigrateFunc is not the new state, but merely a
reference to the old state for error reporting purposes.

The folder state migration behaviour here was set to clone the state to
ensure that the old state was preserved in-flight - we are no longer
doing that as it seems like that is the source of the problem. Now the
state is just modified in place, assuming that refresh will drop a
failed migration's state on the floor.