New resource: vsphere_distributed_port_group
This commit adds the vsphere_distributed_port_group resource, allowing
port groups to be created on DVS switches, introducing the ability to
manage real networks with Terraform that can be used to set up virtual
machines on.

This update also includes various fixes to #188 to fix spurious diffs
and expected behaviour when both resources are used in tandem.