tags: Ensure complete removal of tags is working
tagsClientIfDefined was checking to see if the tags client was defined
as a condition to loading up the tags client for Create and Update
operations. While this generally worked for adding the tags to new
resources and changing tags when necessary, it was not working for
complete tag removal due to its reliance on GetOk, which fails if a
field is removed from config completely.

This fixes tagsClientIfDefined so that it checks both the old and new
incarnations of the tags argument to make sure that both are empty
before passing on loading the tags client.