r/virtual_machine: Allow tuning of customization waiter
The current value is hardcoded at 10 minutes, which has been possibly
causing some issues. This allows for the modification of the timeout -
so that it can be change to be a later value if customization takes a
long time (possible on systems with older hardware or non-SSD disk), or
disabled altogether to eliminate it as a source of problems.

Note that disabling the waiter pushes waiting over to the guest
networking waiter, which is currently also hardcoded at 5 minutes.
Modification of this attribute to a different value will require a state
migration (to translate true => 5min and false => <disabled>) which is a
bit more of an involved change, so I want to take a wait-and-see
approach to see if it is required, especially considering the resource
will be undergoing a larger refactor very soon.