d/distribtued_virtual_switch: Correctly allow unspecified datacenter
As per #203, this fixes the vsphere_distributed_virtual_switch data
source so that it allows for an absolute path in lieu of a datacenter,
as per the specification in the docs.

This was intended, but we were missing the use of GetOk to ensure that
the datacenter was actually specified.