r/virtual_machine: Virtual device CRUD refresh abstraction
The abstraction for virtual devices should now be pretty much complete,
with device-specific CRUD workflows feeding in and out of a more general
core device create/update/delete apply workflow and a read-only refresh
workflow, that also populates devices discovered out-of-band so that
they can be removed later. This has allowed us to develop some better
abstraction of some of the functionality so that all it should really
look like to the higher-level callers (assuming to be the top-level
resource CRUD) is a simple fetching of the current device list at the
time of the apply (or the default device list in the case of a brand
new VM), and then filtering that device list and a list of device
changes through each apply operation, with the final
VirtualDeviceConfigSpec list being used as the deviceChange property.
For reads, the device list can just be fed to the refresh function for
each major device class.

This should set the stage now for us to continue on and finish writing
the CRUD workflow for CDROM devices, which should now be quite simple as
most of the major legwork has been done already.