r/virtual_machine: Basic CRUD complete and testing starting
It's alive! Sorta.

The resource now creates and deletes okay, but there are some
inconsistencies around device management that I need to fix. First off,
on Read, old state cannot be committed of course, so we need to rebuild
the sets from scratch. This means I will really need to figure out a
different way to compute IDs and ensure drift is prevented as much as
possible. Second on the device front is that it looks like the default
device set does not contain unit numbers and as such, it's hard to
determine what unit number a device is actually going to have before
creating the virtual machine.

I will need to think on how to exactly go about this. Again, the
challenges are: 1) trying to keep the whole operation atomic, 2) being
able to uniquely identify devices when there is not much in the way of
stable identifiers to rely on before a device is created, and 3) we need
to be able to prevent unintended drift when specific entry is removed
from configuration (ie: say we have device A, B, and C, and we want to
remove device B - we need to make sure removing device B does not have
unintended consequences for device C when it shits down the list as a
result).

There are also a couple of other options/defaults that need cleaning up
as well.