r/virtual_machine: More tests
Stopping short at modifying existing devices as it turns out that since
I'm using sets with a static hash, modifications to values within the
set does not trigger a diff (like a list would).

This probably means that more than likely we will have to drastically
change how we are doing the VM resource device CRUD, but I want to
commit the functional tests and the changes necessary to fix any bugs
that I encountered along the way before proceeding to any further.