r/virtual_machine: More virtual device stuff, and logs
This update almost stabilizes device workflows, and also adds a bunch of
logging that we were able to use to track down the problems with the
core schema helper's field overwriting logic.

Network interfaces still need a bit of work as it seems that the order
is messed up on update, and I'll be reviewing that shortly.