r/virtual_machine: Add network devices in correct order
For network devices, we infer the order (and the device unit number)
from the order they are added in config, allowing us to skip the
specification of the device unit number. This means that we need to take
care when reading devices in to add them back in the right order, which
was not happening - the order we read in devices (newly added, then
recognized, then orphaned) meant that new devices were possibly taking
the slots that old devices were using.

This has been fixed by setting network interface refresh to add devices
in to the slot of their calculated index - based off of the device's PCI
offset and the current length of the device list.

Also, we have just removed the reference to a temporary map in virtual
devices during new device refresh. The new logic just uses the resource
which is the source of the data anyway.