r/virtual_machine: Complete VM clone spec validation and expansion
This finishes the validation and expansion of the VM clone specs.

Next I want to work on customization spec schema and expansion before I
move on to writing the exact clone workflow. Customization will be set
up in such a way that it could be extended to allow re-customization in
the future.