r/virtual_machine: Import functionality
Import functionality now added.

The import actually survived read very well, which was impressive for
what we are dealing with on the virtual device front. All disks get
added as if they were orphaned, so they get keep_on_remove enabled, but
this, along with other settings that are mainly TF-specific and can't be
read from the VM itself, will be normalized on the next apply safely
(the disk remains intact as CustomizeDiff properly associates the
imported disk with the disk in config during the next plan).

This has an acceptance test present and it has been smoke tested with a
basic config and the import of a manually created VM.