docs: Update provider index page, remove WIP blubs from README and index
Greatly updated the top-level provider documentation so that it's quite
a bit more general. This should stand the test of time now. The VM
resource example from the new version of the resource has also been
copied, which should serve as a good "getting started".

Also removed any notes about things being WIP from both the provider
index and the README on the repo. As we approach 1.0.0, we want to
express that the resource is fully featured enough to be able to be used
day-to-day in production.