docs: More doc updates, deprecate adapter_type in vsphere_virtual_disk
This pretty much completes the standardization of documentation for the
provider, normalizing the various attribute documentation styles into
one, adding various grammar fixes, and notes where they are needed.

Part of this work is deprecating and documenting said deprecation of the
adapter_type parameter of the vsphere_virtual_disk resource. It has no
effect on the type of adapter that the disk gets attached as, and is
removed in future versions of the vSphere disk APIs, which this resource
is not using but may need to in the futrue as the old VirtualDiskManager
MO has been deprecated with vSphere 6.5.