r/virtual_machine: Make sure network changes properly on update
This came up during general use of the pre-release version of the
resource and was missed in testing due to issues with the test template
configuration (the template was set to use the general target network by
default and as such the backing did not need to be updated).

After changing the network on the test template to the more general "VM
Network", was able to be reproduced and fixed so that deployed templates
always have the correct network if they need to be changed.