r/virtual_machine: Add new extraconfig test and up RAM
Added a new extra_config test just to make sure that keys can be
removed. This uncovered a "parameter is incorrect" test, so that was
good.

Also updated all tests to use 2048 MB RAM minimum, as I want to
eventually move the "bare metal" stuff to CoreOS iPXE, which requires at
least 2 GB RAM to run.