docs: Add section on how TF uses managed object IDs
Terraform uses managed object IDs in the vSphere provider to provide a
better interface to downstream resources. This sidesteps a number of
issues that arise from using human-readable values, such as normalizing
different forms of inventory paths, the need to do reverse lookups when
such lookups would be architecturally expensive, and possible issues
dealing with extraneous characters, delegating the logic and possible
issues to data sources as single points of entry.

This adds info in the index on TF's use of MOIDs and also information on
how to fetch MOIDs using govc and the managed object browser.