r/virtual_machine: Add validation for timezone in linux spec
This adds validation for the timezone field in the Linux config spec.

There are actually a good number of combinations in the TZ database, and
some of them don't actually follow the standard "Region/City" format
normally seen. The validation regexp tries to accommodate all cases,
blocking characters that are not currently seen in the VMware API
listing of valid timezones.