docs: Add managed object reference links to VM PR-specific docs
As per the work done in #249, this adds updates to the docs specific to
this PR to link the reader to the general MoRef information now found in
the doc index.

It also links the user to the disk provisioning policy reference at
VMware so that one can get a full explanation on what each disk
provisioning option means, ie: what is the difference between thin,
eager zeroed, lazy zeroed, etc.