r/virtual_machine: Roll back VMs on customization errors
Errors in either sending the customization spec, or waiting for
customization (if the customization waiter is set to do so) should roll
back in the same way as reconfigure does. This is because there might
not necessarily be an issue with the user's customization options -
there could be issues with the base VM, or some other issue that results
in a situation where the correct config is sent to state, which will
require a manual taint or destroy/re-create to correct without this
behaviour.