r/virtual_machine: Don't delete VM on customization waiter errors
Rolling back part of the previous commit where the rollbacks were added.
When customization errors happen, there is often valuable info in the
customization logs on the created VM that can assist with customization
troubleshooting. These will be deleted if the VM is rolled back.

Instead, we provide a verbose error with the error from the event and a
referral to the "terraform taint" command reference so that a user can
check their error, correct it, and then taint the resource manually, if
no modification to the actual customization configuration was necessary.

Also added extra info on tainting to the post-clone rollback message, as
the VM more than likely will actually exist in state. Also removed the
"dangling resource" part of the warning, as the resource is technically
in state and not dangling.