r/virtual_machine: Add default for disk unit number
This fixes a diff mismatch on tainting of a resource that may possibly
come up during ForceNew as well. The annoying part of this issue is that
the "diffs didn't match during apply" error fires during the middle of
the operation, after the resource gets deleted, but before it's
re-created.