r/host_virtual_switch: Support a switch with no NICs
This is basically done in vSphere by supplying no bridge to the config
spec - in our configuration, we were passing the bridge with zero nics,
which is an invalid configuration.

This change just nils out the bridge after expansion if there are no
NICs. We have also added some diff-time validation to make sure that any
NIC specified in the active or standby list is a NIC configured on the
bridge to avoid errors at apply time.