r/virtual_machine: Fix folder issues
Some issues with the folder location code, which was restricted to
mainly vCenter things before, is now being used by the VM resource,
which of course needs to be supported on standalone ESXi as well. This
was not happening as we were gating full access to most of the folder
code to vCenter.

This allows the code to work as long as no folder paths are specified.
It also adds a small exception for checking folder types when the child
type is []{"VirtualMachine"}, instead of []{"Folder", "VirtualMachine"},
which is how it is on standalone ESXi, as you cannot add folders to the
root folder in the "inventory".