r/virtual_machine: Device cull range fix, and handle zero-NIC templates
There was some issues with our device culling/ranging logic in certain
scenarios:

* On templates with no NICs, unitRange was not producing a proper
result. This was the result of these values being pointers because we
work with the unit number value directly in the virtual device, which is
a pointer. This fixes things so that if a zero-length device list is
passed to unit range, we return 0 immediately (as there are no devices
to range over).
* On templates where the NIC count was less than the amount of NICs
needed by configuration, the device cull range expression (which is used
to handle the exact opposite scenario) was running into out-of-range
issues. This has been fixed by asserting that the old device list is
indeed longer than the new device list so that there's no way that
referencing old[len(new):] will fail. This has been extended to CD
devices too, which has the same logic, even though it's currently masked
by the fact that we only allow one cdrom device.