viapi: Fix version comparison logic
Version comparison logic was not sound across minor and patch versions.
The new logic just bit-shifts major and minor so that a nice clean
comparable integer. I'd imagine this will be fine, as I've never seen
VMware put out any sort of release with a version anywhere close to the
8-bit space we give each version particle now.

Fixes #265.