r/virtual_machine: Validate exact disk type when using linked_clone
A VM that's cloned with linked_clone and different disk mode options
that what is set on the source template will end up with inconsistent
settings post-clone, causing a read-update failure that will leave the
VM in a state that is irrecoverable to Terrform - the only way to
rectify the situation is to delete the VM and start over with a new
state.

This adds extra validation to make sure that thin_provisioned and
eagerly_scrub are set to the same settings as the source template's
counterpart disk, along with validating that the disk is the exact same
size (something that wasn't done yet).

Fixes #275.