d/virtual_machine: Add extra disk attributes
This moves disk attributes to the disks sub-resource, which now fetches
the settings of size, eagerly_scrub, and thin_provisioned, as all 3 need
to be the same when doing linked clones.

This covers the data source - however, the counterpart resource
documentation and tests need to be fixed to use the new structure.