r/virtual_machine: Correct present NIC device calculation
This corrects an issue with the way NIC devices were checked to see if
the slots they are supposed to occupy on the PCI bus were already used
or not. An offset that was used to populate the current device range was
not being used when the devices were actually checked, creating invalid
results and out of range errors.

Fixes #279.