r/virtual_machine: Adjust tests to use new VM data source
This adds the extra options to use for linked cloning and otherwise
modifies the consumption of the data source to the new, correct syntax.

This also adds extra error cases for bad size and
linked_clone/eagerly_scrub options to test for discrepancies in these
options in both linked and unlinked scenarios.