Remove linked clone, update docs
Using a linked clone is not correct for this example as we will be
cloning to a different datastore than the template resides in, more than
likely. Further to this, we want to be able to demonstrate storage
vMotion in our blog post, which will not be possible with a linked
clone.