r/virtual_machine: Failsafe against disk name mismatch
There are certain situations that can lead to a disk ultimately losing
its name at a specific location without Terraform being able to keep
track of it:

* Significant alteration of a VM config outside of Terraform: It's
feasible that someone could go and completely replace a virtual disk in
one location with a different one.
* Snapshot heck: Snapshots alter the name of a virtual disk. We walk up
the child disk path to ultimately try and confirm if we are dealing with
the right virtual disk, but it's not out of the realm of possibility
that with enough manipulation, the original name of the disk might be
lost.
* Linked clones and storage vMotion: This is currently a real
possibility and the main reason we are putting in this validation.
vMotions of linked clones ultimately assign delta disk-like names to
virtual disks, completely removing the original name from the chain.

In these situations, the existing state of the disk does not facilitate
adding keep_on_remove (unless explicitly there already) as the logic
does not believe the disk to be orphaned (it's in state and was found in
config by key or device address). This can lead to potentially
destructive scenarios.

This update guards against that - if the base name of the disk that was
in state before (if any) and the disk ultimately read out of the
backing do not match, taking snapshots into account, an error is kicked
back during the read process and TF stops in its tracks. Any state that
is pending from DiskRefreshOperation is not committed, and an error
message is supplied prompting the user to manually correct the issues
with the VM before (carefully) proceeding again.

Further commits will close the hole in the final scenario, preventing
storage vMotion on linked clones (ie: the root disk in the snapshot
chain does not match the file name of the disk we are managing). This
should prevent this error from coming up during normal TF operations.