r/virtual_machine: Block VM renames that result in disk deletion
This is another fail-safe - this time, we are blocking virtual machine
rename operations that coincide with the deletion of any disks that bear
the name.

This prevents a scenario where disk names are being controlled by the
same variable that controls the virtual machine name - a very likely
scenario, given that clones require that the disks bear the name of the
virtual machine.

Also, documentation for the virtual machine resource has been updated
with a suggested workaround, using null-resource with an ignore_changes
to store the original name permanently.