r/virtual_machine: Add test case for orphaned disk on mismatch
Lots of helper code added in this test to help manage datastore paths
and what not - things we are kind of doing in other places manually but
I'm getting tired of writing out. Also, a helper for moving a virtual
disk.

The main test case fails right now which is intentional, it fails
correctly which is the good thing. Again, this is testing for the case
describe in #304 where if a disk with a different name gets put into a
place where TF already is tracking, that disk will get deleted on the
next TF run. We are working to fix it so that it simply removes it.

Ref: #304.