r/virtual_machine: Do not delete disks at same address as existing disk
There are certain situations where the filename on a disk can get lost,
namely associated with snapshots, and storage vMotion on VMs that have
been created from linked clones or when a virtual machine has been
renamed.

In light of this, long-term we are planning on transitioning away from
being able to control the name of virtual disks that are not attached
externally. This stop-gap ensures that the disk is kept in situations
like this - where we actually find a disk recognized at a specific
location that we are tracking, versus just deleting the disk because the
logic does not consider it "orphaned".