r/virtual_machine: Document current storage vMotion restrictions
All of these restrictions will more than likely be removed or modified
in some way. The external attachment restriction will be a challenge,
but ultimately we want to try and control the lifecycle of that outside
of the VM, and work to re-attach that disk somehow after the migration.
How we go about that remains to be seen.