docs: Clarify the use of cluster/host root resource pool
There has been some confusion surrounding the fact that the
vsphere_virtual_machine resource now only takes a resource pool ID with
users that do not make use of resource pools, opting to just put virtual
machines in the root-level cluster or standalone host. This actually
places those virtual machines in a root resource pool that represents
all of the compute power of a particular host or set of hosts.  This
pool is normally invisible to the user and can only be seen generally by
using a tool that can browse the inventory off the CLI, like govc.

There are no plans to complicate the VM resource's implementation by
adding parameters like compute_resource_id - there is no point as they
would ultimately be looking up a resource pool anyway. Future plans may
see the development of data sources like vsphere_cluster and
vsphere_standalone_host, which could possibly have resource_pool_id
exported which would reference the root resource pool for that specific
resource. These may be developed along with their resource counterparts,
which are planned for a future release.

To back up the above, this commit adds a bit of clarification in the VM
resource on why only a resource pool is accepted - this can be expanded
on after we add the cluster and standalone host data sources. I've also
rephrased the explanation in the vsphere_resource_pool documentation so
that it's bit more clear what the root resource pool is and that this
logic also applies to standalone hosts. Also, the language has been
changed from using "default" to using "root", so that it's in line with
the vSphere official docs.