r/virtual_machine: Fix NIC device offset misalignment crashes
Looks like unitRange, a virtual device (read: virtual NIC device) helper
that helps calculate the device unit range we need to account for when
refreshing devices, and also running post-clone device modification
operations, was not behaving as intended. We were incorrectly using the
current NIC set as the source of truth for the unit floor, in addition
to correctly using it for the device ceiling). This was creating issues
where the current NIC set started at an offset that was higher than
where NIC devices start on the PCI bus.

This fix corrects that - the NIC PCI device offset constant is now what
defines the device floor, and is immutable. The only thing that gets
modified is the ceiling. Added a few unit tests to back it up.

Also renamed the function nicUnitRange to reflect the fact that no other
device type is using this logic right now.

Fixes #341.